Canonicalise source of ruby version
Most Ruby version managers respect the `.ruby-version` file and will
auto-switch your local Ruby version to the version this file specifies
in development (or indeed any other environment). This ensures you are
always running the expected Ruby interpreter.

The trigger for this PR was my need to explicitly specify the local Ruby
version when setting up the project today. So while quite specific to my
needs - overall it is hopefully not too contentious a change and may
simplify future onboarding.