Allow hidden locations to be returned in locations.json
Add a flag to allow hidden locations to be included
in locations.json. This is required to get all locations
for the data platform.

Note: This is specifically to ensure we get Caerphilly
in the feed which is a hidden location as it does 
not have appointments, it is a call centre only.