Allow edits to `online_booking_twilio_number`
Grants administrators the ability to enter the online booking specific
twilio number on new or existing locations.

The number is now nullable, this is since nulls with get through when
the params are filtered (or when it isn't supplied) whereas previously
this was something that was only added through the console.