Permit first coordinates if there are multiple
This isn't entirely fool proof but rather than stating these are
invalid and not permitting the user to proceed, we can just accept the
first set of coordinates and move along. In practice the actual gaps
between coordinates are so trivially small that picking the first is
usually the right thing to do.