Allow children to opt out of online booking
Online bookings for locations was previously controlled by the booking
location or parent location being enabled for online bookings. This
change allows us to specify that the booking location itself is enabled
for booking, but also whether individual children are enabled.

This change required a migration to blanket apply the
`online_booking_enabled` flag to existing child locations with booking
locations also enabled for online bookings already since previously the
flag was inferred from the parent as previously stated.