Include hidden / inactive locations in response
We need to include hidden locations in the booking location API response
to insulate against changes being made to the size and structure of
booking locations and their respective child locations.

This isn't such an issue when used from the frontend's perspective but
causes lots of issues when planner refers to locations that no longer
'exist' in the API-sense.