Begin conversion of `guiders` association
We're converting the `guiders` association from `has_many` to
`has_many through:` to allow us to assign guiders to multiple
locations. This is to support the reassignment of child locations to
booking locations.

When a child moves to a new parent it needs to bring with it the
associated guiders for the parent location to maintain data integrity
via the booking locations API.

This processed is stepped as we are migrating from a typical
`belongs_to` to a join table approach and this cannot be migrated in a
single atomic step.