Introduce a temporary cut-off period
When we specify a `cut_off_to` date, slots will not be available for
bookings until after that date. Unlike `cut_off_from` - a hard closure
of a location - this still keeps the location open and available to take
bookings but introduces a 'breathing' period to ease processing a
backlog of bookings.