Allow bookings during bank holidays
Before this change we hid the online booking button. We now allow
customers to enter the bookings process regardless whether the current
day is a public holiday. This makes sense as the slots and cut-off
period logic will not allow a customer to book an appointment *for* a
public holiday, yet nothing ought to stop a customer from placing a
booking during this period.